Technically kind of working, but not really.
The edge cases aren't really accounted for (when bumping up against the
edge, which is when next_index < 0 or >= len of siblings).

There is also a case where moving to the left from a subcontainer into
where a view used to be as an ancestor puts it in the wrong place
(left side not right side).

Going to fix that last one now